fix(geometries): Fix stringifyFloat to correctly format number
We missed the "." here and thus PHP formatted numbers with the default precision of 6 and prefixed it with whitespaces.

Additionally the trims were adjusted, as we now do not need to trim any whitespaces. Also the trimming for the "." was made a rtrim instead to avoid wrongly stripping `.1` to `1`, even though this should never be a possible value that sprintf would produce.

See PHP docs: https://www.php.net/manual/en/function.sprintf.php

Thanks @ronnypolloqueri for reporting this.